+++ src/components/Kebabmenu.js
@@ -1,41 +1,14 @@
 import { useState } from "react";
 import { createPortal } from "react-dom";
 
-import styled from "styled-components";
-import kebabIcon from "assets/kebab.svg";
 import ModalContainer from "./modal/ModalContainer";
 import AddFolder from "./modal/AddFolder";
 import DeleteFolder from "./modal/DeleteFolder";
+import PopoverMenu from "./modal/PopoverMenu";
 
-const PopoverMenu = styled.div`
-  position: absolute;
-  top: 20px;
-  right: 0;
-  width: 6.25rem;
-  height: 4rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 4px;
-  background: var(--color-white);
-  box-shadow: 0px 2px 8px 0px rgba(51, 50, 54, 0.1);
-`;
+import kebabIcon from "assets/kebab.svg";
 
-const DeleteOption = styled.button`
-  padding: 7px 12px;
-  font-size: 0.875rem;
-  color: var(--color-text);
-  background: none;
-  border: none;
-  cursor: pointer;
-`;
-
-const AddOption = styled(DeleteOption)`
-  color: var(--color-primary);
-  background: var(--color-primary-bg);
-`;
-
-export default function KebabMenu() {
+export default function KebabMenu({ link }) {
   const [isOpenMenu, setIsOpenMenu] = useState(false);
   const [iseOpenModal, setIsOpenModal] = useState(false);
   const [selectedOption, setSelectedOption] = useState("");
@@ -54,12 +27,7 @@ export default function KebabMenu() {
   return (
     <>
       <img src={kebabIcon} onClick={handleMenuToggle} alt="memu" />
-      {isOpenMenu && (
-        <PopoverMenu onClick={handleModalOpen}>
-          <DeleteOption id="deleteLink">삭제하기</DeleteOption>
-          <AddOption id="addLink">폴더에 추가</AddOption>
-        </PopoverMenu>
-      )}
+      {isOpenMenu && <PopoverMenu onClickModal={handleModalOpen} />}
 
       {iseOpenModal &&
         createPortal(
@@ -67,7 +35,7 @@ export default function KebabMenu() {
             {selectedOption === "addLink" ? (
               <AddFolder />
             ) : (
-              <DeleteFolder currentFolderName="test" />
+              <DeleteFolder currentFolderName={link} />
             )}
           </ModalContainer>,
           document.getElementById("portal"),
